backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyyaml==6.0.2
pydantic==2.10.4
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.6
opencv-python-headless==4.10.0.84
httpx==0.28.1
numpy==2.2.1
minio==7.2.15
pytest==8.3.4
pytest-asyncio==0.25.0
